% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{filtrar_por_vocal}
\alias{filtrar_por_vocal}
\title{Filtro según última letra}
\usage{
filtrar_por_vocal(palabras)
}
\arguments{
\item{palabras}{Vector de palabras}
}
\value{
Subconjunto de palabras que terminan en vocal
}
\description{
Esta función recibe un vector de palabras y devuelve sólo aquellas que
terminan en vocal.
}
\details{
Si \code{palabras} no es un vector carácter, se produce un error.
}
\examples{
x <- c("casa", "flor", "pared", "día", "sol")
filtrar_por_vocal(x)
filtrar_por_vocal(c("montes", "cuadernos"))

}
